requirements.txt
httplib2
Flask
flask-swagger
pillow
numpy
pytest
ultralytics
SQLAlchemy
firebase-auth
transformers